rename stosk module to stocks
